+++ packages/pasteboard/CHANGELOG.md
@@ -1,6 +1,11 @@
+## 0.1.1
+
+* add `html` to get the html from clipboard. [@alexkmbk](https://github.com/alexkmbk)
+* support write image on web.
+
 ## 0.1.0
 
-* add web supoort.
+* add web support.
 * add `writeText` and `text` support.
 
 ## 0.0.3
@@ -14,4 +19,4 @@
 ## 0.0.1
 
 * add Linux,Windows,macOS support.
-* suport basic functions.
+* support basic functions.